// frontend/src/api/generated/model/engagementCardCardType.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * HuaXia Tourism RAG
 * OpenAPI spec version: 0.1.0
 */

export type EngagementCardCardType = typeof EngagementCardCardType[keyof typeof EngagementCardCardType];


export const EngagementCardCardType = {
  attraction_knowledge: 'attraction_knowledge',
  city_folk_custom: 'city_folk_custom',
  local_flavor: 'local_flavor',
  traveler_reminder: 'traveler_reminder',
} as const;